Struct spaceapi_server::sensors::SensorSpec

pub struct SensorSpec {
    pub template: Box<SensorTemplate>,
    pub data_key: String,
}

A specification of a sensor.

The template field contains the static data of a sensor and the data_key says how to find the sensor value in Redis.

Fields
template: Box<SensorTemplate> data_key: String
Methods

impl SensorSpec

fn get_sensor_value(
&self,
redis_pool: Pool<RedisConnectionManager>
) -> Result<String, SensorError>

Retrieve sensor value from Redis.

fn set_sensor_value(
&self,
redis_pool: Pool<RedisConnectionManager>,
value: &str
) -> Result<(), SensorError>

Set sensor value in Redis.