Struct spaceapi_server::SpaceapiServer

pub struct SpaceapiServer { /* fields omitted */ }

A Space API server instance.

You can create a new instance using the new constructor method by passing it the host, the port, the Status object and a redis connection info object.

The SpaceapiServer includes a web server through Hyper. Simply call the serve method.

Methods

impl SpaceapiServer

fn serve<S: ToSocketAddrs>(self, socket_addr: S) -> HttpResult<Listening>

Start a HTTP server listening on self.host:self.port.

The call returns an HttpResult<Listening> object, see http://ironframework.io/doc/hyper/server/struct.Listening.html for more information.