Struct spaceapi_server::sensors::SensorSpec

pub struct SensorSpec {
    pub template: Box<SensorTemplate>,
    pub data_key: String,
}

A specification of a sensor.

The template field contains the static data of a sensor and the data_key says how to find the sensor value in the datastore.

Fields

template

data_key

Methods

impl SensorSpec

fn get_sensor_value(&self, datastore: SafeDataStore) -> Option<String>

Retrieve sensor value from the datastore.